requirements.txt
praw
requests
beautifulsoup4
pandas
click
rich
tqdm
pyyaml
lxml
python-dateutil